// packages/core/src/games/pong/types.ts
import { Component, CoreComponentRegistry } from "../../index";

export interface BallComponent extends Component {
  type: "Ball";
  spinFactor: number;
  spinDecay: number;
  visibilityTimer?: number;
}

export interface PaddleComponent extends Component {
  type: "Paddle";
  side: "left" | "right";
  previousY: number;
  lastVelocityY: number;
}

export interface PongState extends Component {
  type: "PongState";
  scoreP1: number;
  scoreP2: number;
  isGameOver: boolean;
  winner?: number;
  comboMultiplier: number;
  gameOverLogged?: boolean;
}

export interface PongInput extends Record<string, boolean> {
  p1Up: boolean;
  p1Down: boolean;
  p2Up: boolean;
  p2Down: boolean;
}

export interface PongInputFrame {
  tick: number;
  input: PongInput;
}

export interface PongRoomState {
  gameStarted: boolean;
  serverTick: number;
  p1Connected: boolean;
  p2Connected: boolean;
  seed: number;
}

export interface PongComponentRegistry extends CoreComponentRegistry {
  Ball: BallComponent;
  Paddle: PaddleComponent;
  PongState: PongState;
}

export const PONG_CONFIG = {
  WIDTH: 800,
  HEIGHT: 600,
  PADDLE_WIDTH: 15,
  PADDLE_HEIGHT: 80,
  BALL_SIZE: 10,
  PADDLE_SPEED: 400,
  BALL_SPEED_START: 300,
  BALL_SPEED_INC: 1.05,
  WIN_SCORE: 5,
  BALL_INVISIBLE_AFTER_HIT_TICKS: 0,
};